requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.9.1.post1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.20
botocore==1.29.20
certifi==2022.9.24
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
common==0.1.2
confluent-kafka==2.2.0
decorator==5.1.1
Django==4.0.3
django-session-timeout==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.4.0
elasticsearch==8.5.2
et-xmlfile==1.1.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.23.5
openai==0.27.3
openpyxl==3.0.10
pandas==1.5.2
psycopg2-binary==2.9.5
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rest-framework-simplejwt==0.0.2
s3transfer==0.6.0
self==2020.12.3
six==1.16.0
sqlparse==0.4.3
tanium==0.0.dev1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
Werkzeug==2.2.3
yarl==1.8.2